pyproject.toml
[project]
name = "oss-sanitizer"
version = "0.1.0"
description = "Source code compliance scanner for the Geneva Open Source Charter"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "gitpython>=3.1",
    "openai>=1.0",
    "anthropic>=0.40",
    "pyyaml>=6.0",
    "rich>=13.0",
]

[project.scripts]
oss-sanitizer = "oss_sanitizer.cli:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/oss_sanitizer"]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
